[Proyect] Add: Mejoras en el rendimiento y uso del componente de TableData, ahora puede trabjar con cualquier información que se le pase como argumentos en la  ruta onde se vaya a utilizar, se ha implementado ya en History y Match del administrador
