+++ FrontEnd/src/pages/Admin/Matches/Matches.css
@@ -1,13 +1,13 @@
-.admin-matches-page-containe{
-    width: 100%;
-    height: auto;
-    display: grid;
-    grid-template-columns: repeat(8, 1fr);
+.admin-matches-page-containe {
+  width: 100%;
+  height: auto;
+  display: grid;
+  grid-template-columns: repeat(8, 1fr);
 }
 
 .admin-matches-view-content {
-    grid-column: 1/-1;
-    min-height: 50vh;
-    width: 90%;
-    margin: 0 auto 5rem;
+  grid-column: 1/-1;
+  min-height: 50vh;
+  width: 90%;
+  margin: 2rem auto 5rem;
 }
